# backend/specs/1-rag-chatbot-backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: RAG Chatbot Backend API
  description: Retrieval-Augmented Generation chatbot for Physical AI & Humanoid Robotics textbook
  version: 1.0.0
  contact:
    name: Backend Engineering Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production
paths:
  /chat/query:
    post:
      summary: Submit query to RAG chatbot (full-book search)
      operationId: submitQuery
      tags:
        - Chat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
            examples:
              student_question:
                summary: Student asking about robot control
                value:
                  query: "What is the difference between hierarchical and decentralized robot control?"
              educator_fact_check:
                summary: Educator verifying fact
                value:
                  query: "What page discusses passive compliance in robotics?"
      responses:
        '200':
          description: Query answered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
              examples:
                success:
                  summary: Successful answer with citations
                  value:
                    answer_id: "550e8400-e29b-41d4-a716-446655440000"
                    answer: "Hierarchical robot control uses a centralized decision-maker [Source: Control Architectures (Page 89) - \"hierarchical systems employ a central controller that coordinates all subsystems\"] while decentralized control distributes decision-making across nodes [Source: Distributed Control Paradigms (Page 102) - \"decentralized approaches empower individual agents to make local decisions\"]."
                    sources:
                      - id: "chunk-001"
                        excerpt: "hierarchical systems employ a central controller that coordinates all subsystems"
                        page: 89
                        section: "Control Architectures"
                      - id: "chunk-002"
                        excerpt: "decentralized approaches empower individual agents to make local decisions"
                        page: 102
                        section: "Distributed Control Paradigms"
                    confidence: 0.93
                    confidence_level: "high"
                    status: "success"
        '400':
          description: Invalid query format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                empty_query:
                  summary: Empty query
                  value:
                    error:
                      code: "INVALID_QUERY"
                      message: "Query must be 1-1000 characters"
                      details: "Provided: empty string (0 characters)"
                query_too_long:
                  summary: Query exceeds 1000 characters
                  value:
                    error:
                      code: "INVALID_QUERY"
                      message: "Query must be 1-1000 characters"
                      details: "Provided: 1200 characters"
        '401':
          description: Unauthorized (invalid API key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Query timeout (exceeded 5 seconds)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                timeout:
                  summary: Query processing timeout
                  value:
                    error:
                      code: "TIMEOUT"
                      message: "Query processing timed out"
                      details: "Exceeded 5 second limit. Try a simpler question."
...
components:
  schemas:
...
    QueryRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 1000
          description: User question in natural language
          example: "What is passive compliance in robotics?"
        conversation_id:
          type: string
          format: uuid
          description: Optional conversation ID for multi-turn support (future)
        mode:
          type: string
          enum: [full-book]
          default: full-book
          description: Retrieval mode (always full-book for this endpoint)
...
    QueryResponse:
      type: object
      required:
        - answer_id
        - answer
        - sources
        - confidence
        - status
      properties:
        answer_id:
          type: string
          format: uuid
          description: Unique answer identifier
        answer:
          type: string
          description: Generated response with inline citations
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          description: List of cited sources
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Confidence score (0.0-1.0)
        confidence_level:
          type: string
          enum: [high, medium, low]
          description: Qualitative confidence level
        status:
          type: string
          enum: [success, partial, out_of_scope, error]
          description: Answer status
...
    Source:
      type: object
      required:
        - id
        - excerpt
        - page
        - section
      properties:
        id:
          type: string
          description: Unique chunk identifier
        excerpt:
          type: string
          description: Relevant passage from source
        page:
          type: integer
          description: Page number where excerpt appears
        section:
          type: string
          description: Section heading for navigation
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_QUERY
                - LOW_CONFIDENCE
                - TIMEOUT
                - NO_MATCH
                - PASSAGE_INSUFFICIENT
                - UNAUTHORIZED
                - SERVER_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: string
              description: Additional context for debugging
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as bearer token
...